refactoring and cleaning
