+++ print3/config.py
@@ -1,0 +1,8 @@
+
+MAPFISH_FILE_PREFIX = 'mapfish-print'
+MAPFISH_MULTI_FILE_PREFIX = MAPFISH_FILE_PREFIX + '-multi'
+USE_MULTIPROCESS = True
+VERIFY_SSL = False
+LOG_SPEC_FILES = False
+REFERER_URL = 'https://map.geo.admin.ch'
+USE_LV95_SERVICES = False
