--- print3/tests/integration/test_checker.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from print3.tests.integration import TestsBase
-
-
-class TestCheckerView(TestsBase):
-
-    def test_checker(self):
-        self.testapp.get('/checker', params={'url': 'http://s.geo.admin.ch/e83c57af1'}, status=200)
-
-    def test_checker_dev(self):
-        self.testapp.get('/checker_dev', params={'url': 'http://s.geo.admin.ch/e83c57af1'}, status=200)
