--- print3/tests/manual/print/movie.json
@@ -1,265 +1,0 @@
-{
-  "layout": "1 A4 landscape",
-  "srs": "EPSG:21781",
-  "units": "m",
-  "rotation": 0,
-  "app": "config",
-  "lang": "it",
-  "dpi": "150",
-  "layers": [
-    {
-      "layer": "ch.swisstopo.zeitreihen",
-      "opacity": 1,
-      "type": "WMTS",
-      "baseURL": "http://wmts.geo.admin.ch",
-      "maxExtent": [
-        420000,
-        30000,
-        900000,
-        350000
-      ],
-      "tileOrigin": [
-        420000,
-        350000
-      ],
-      "tileSize": [
-        256,
-        256
-      ],
-      "resolutions": [
-        4000,
-        3750,
-        3500,
-        3250,
-        3000,
-        2750,
-        2500,
-        2250,
-        2000,
-        1750,
-        1500,
-        1250,
-        1000,
-        750,
-        650,
-        500,
-        250,
-        100,
-        50,
-        20,
-        10,
-        5,
-        2.5,
-        2
-      ],
-      "zoomOffset": 0,
-      "version": "1.0.0",
-      "requestEncoding": "REST",
-      "formatSuffix": "png",
-      "style": "default",
-      "dimensions": [
-        "TIME"
-      ],
-      "params": {
-        "TIME": "20121231"
-      },
-      "matrixSet": "21781",
-      "timestamps": [
-        "20121231",
-        "20111231",
-        "20101231",
-        "20091231",
-        "20081231",
-        "20071231",
-        "20061231",
-        "20051231",
-        "20041231",
-        "20031231",
-        "20021231",
-        "20011231",
-        "20001231",
-        "19991231",
-        "19981231",
-        "19971231",
-        "19961231",
-        "19951231",
-        "19941231",
-        "19931231",
-        "19921231",
-        "19911231",
-        "19901231",
-        "19891231",
-        "19881231",
-        "19871231",
-        "19861231",
-        "19851231",
-        "19841231",
-        "19831231",
-        "19821231",
-        "19811231",
-        "19801231",
-        "19791231",
-        "19781231",
-        "19771231",
-        "19761231",
-        "19751231",
-        "19741231",
-        "19731231",
-        "19721231",
-        "19711231",
-        "19701231",
-        "19691231",
-        "19681231",
-        "19671231",
-        "19661231",
-        "19651231",
-        "19641231",
-        "19631231",
-        "19621231",
-        "19611231",
-        "19601231",
-        "19591231",
-        "19581231",
-        "19571231",
-        "19561231",
-        "19551231",
-        "19541231",
-        "19531231",
-        "19521231",
-        "19511231",
-        "19501231",
-        "19491231",
-        "19481231",
-        "19471231",
-        "19461231",
-        "19451231",
-        "19441231",
-        "19431231",
-        "19421231",
-        "19411231",
-        "19401231",
-        "19391231",
-        "19381231",
-        "19371231",
-        "19361231",
-        "19351231",
-        "19341231",
-        "19331231",
-        "19321231",
-        "19311231",
-        "19301231",
-        "19291231",
-        "19281231",
-        "19271231",
-        "19261231",
-        "19251231",
-        "19241231",
-        "19231231",
-        "19221231",
-        "19211231",
-        "19201231",
-        "19191231",
-        "19181231",
-        "19171231",
-        "19161231",
-        "19151231",
-        "19141231",
-        "19131231",
-        "19121231",
-        "19111231",
-        "19101231",
-        "19091231",
-        "19081231",
-        "19071231",
-        "19061231",
-        "19051231",
-        "19041231",
-        "19031231",
-        "19021231",
-        "19011231",
-        "19001231",
-        "18991231",
-        "18981231",
-        "18971231",
-        "18961231",
-        "18951231",
-        "18941231",
-        "18931231",
-        "18921231",
-        "18911231",
-        "18901231",
-        "18891231",
-        "18881231",
-        "18871231",
-        "18861231",
-        "18851231",
-        "18841231",
-        "18831231",
-        "18821231",
-        "18811231",
-        "18801231",
-        "18791231",
-        "18781231",
-        "18771231",
-        "18761231",
-        "18751231",
-        "18741231",
-        "18731231",
-        "18721231",
-        "18711231",
-        "18701231",
-        "18691231",
-        "18681231",
-        "18671231",
-        "18661231",
-        "18651231",
-        "18641231",
-        "18631231",
-        "18621231",
-        "18611231",
-        "18601231",
-        "18591231",
-        "18581231",
-        "18571231",
-        "18561231",
-        "18551231",
-        "18541231",
-        "18531231",
-        "18521231",
-        "18511231",
-        "18501231",
-        "18491231",
-        "18481231",
-        "18471231",
-        "18461231",
-        "18451231",
-        "18441231"
-      ]
-    }
-  ],
-  "qrcodeurl": "http://api3.geo.admin.ch/qrcodegenerator?url=http%3A%2F%2Fmap.geo.admin.ch%2F%3FX%3D199810.965%26Y%3D600854.95%26zoom%3D5%26lang%3Dit%26topic%3Dech%26bgLayer%3Dch.swisstopo.pixelkarte-farbe%26layers_timestamp%3D20121231%26time%3D2012%26layers%3Dch.swisstopo.zeitreihen",
-  "movie": true,
-  "pages": [
-    {
-      "center": [
-        600854.946041163,
-        199810.959191038
-      ],
-      "bbox": [
-        593781.751596719,
-        204485.26474659427,
-        607928.140485608,
-        195136.65363548315
-      ],
-      "display": [
-        802,
-        530
-      ],
-      "scale": "50000.0",
-      "dataOwner": "© swisstopo",
-      "shortLink": "http://mf-chsdi3.dev.bgdi.ch/mom_print_timestamp/60fdd1efc1",
-      "rotation": 0,
-      "langit": true
-    }
-  ]
-}
