--- print3/tests/manual/print/non-movie.json
@@ -1,156 +1,0 @@
-{
-  "layout": "1 A4 landscape",
-  "srs": "EPSG:21781",
-  "units": "m",
-  "rotation": 0,
-  "app": "config",
-  "lang": "it",
-  "dpi": "150",
-  "layers": [
-    {
-      "layer": "ch.swisstopo.pixelkarte-farbe",
-      "opacity": 1,
-      "type": "WMTS",
-      "baseURL": "http://wmts.geo.admin.ch",
-      "maxExtent": [
-        420000,
-        30000,
-        900000,
-        350000
-      ],
-      "tileOrigin": [
-        420000,
-        350000
-      ],
-      "tileSize": [
-        256,
-        256
-      ],
-      "resolutions": [
-        4000,
-        3750,
-        3500,
-        3250,
-        3000,
-        2750,
-        2500,
-        2250,
-        2000,
-        1750,
-        1500,
-        1250,
-        1000,
-        750,
-        650,
-        500,
-        250,
-        100,
-        50,
-        20,
-        10,
-        5,
-        2.5,
-        2,
-        1.5,
-        1,
-        0.5
-      ],
-      "zoomOffset": 0,
-      "version": "1.0.0",
-      "requestEncoding": "REST",
-      "formatSuffix": "jpeg",
-      "style": "default",
-      "dimensions": [
-        "TIME"
-      ],
-      "params": {
-        "TIME": "20140520"
-      },
-      "matrixSet": "21781"
-    },
-    {
-      "layer": "ch.swisstopo.zeitreihen",
-      "opacity": 1,
-      "type": "WMTS",
-      "baseURL": "http://wmts.geo.admin.ch",
-      "maxExtent": [
-        420000,
-        30000,
-        900000,
-        350000
-      ],
-      "tileOrigin": [
-        420000,
-        350000
-      ],
-      "tileSize": [
-        256,
-        256
-      ],
-      "resolutions": [
-        4000,
-        3750,
-        3500,
-        3250,
-        3000,
-        2750,
-        2500,
-        2250,
-        2000,
-        1750,
-        1500,
-        1250,
-        1000,
-        750,
-        650,
-        500,
-        250,
-        100,
-        50,
-        20,
-        10,
-        5,
-        2.5,
-        2
-      ],
-      "zoomOffset": 0,
-      "version": "1.0.0",
-      "requestEncoding": "REST",
-      "formatSuffix": "png",
-      "style": "default",
-      "dimensions": [
-        "TIME"
-      ],
-      "params": {
-        "TIME": "20121231"
-      },
-      "matrixSet": "21781"
-    }
-  ],
-  "qrcodeurl": "http://mf-chsdi3.dev.bgdi.ch/mom_print_timestamp/qrcodegenerator?url=http%3A%2F%2Fmf-geoadmin3.dev.bgdi.ch%2Fltmom%2Fsrc%2F%3FX%3D199810.96%26Y%3D600854.95%26zoom%3D5%26lang%3Dit%26topic%3Ddev%26bgLayer%3Dch.swisstopo.pixelkarte-farbe%26layers_timestamp%3D20121231%26catalogNodes%3D9900%26layers%3Dch.swisstopo.zeitreihen%26time%3D2012",
-  "movie": false,
-  "pages": [
-    {
-      "center": [
-        600854.9460411635,
-        199810.9591910387
-      ],
-      "bbox": [
-        593781.751596719,
-        204485.26474659427,
-        607928.140485608,
-        195136.65363548315
-      ],
-      "display": [
-        802,
-        530
-      ],
-      "scale": "50000.0",
-      "dataOwner": "© swisstopo",
-      "shortLink": "http://mf-chsdi3.dev.bgdi.ch/mom_print_timestamp/60fdd1efc1",
-      "rotation": 0,
-      "langit": true,
-      "timestamp": "2012"
-    }
-  ]
-}
