--- production.ini.in
@@ -1,80 +1,0 @@
-###
-# app configuration
-# http://docs.pylonsproject.org/projects/pyramid/en/latest/narr/environment.html
-###
-
-[app:main]
-use = egg:print3
-app_version = ${app_version}
-server_id = ${server_id}
-entry_path = ${apache_entry_path}
-available_languages = de fr it en rm
-
-pyramid.reload_templates = false
-pyramid.debug_authorization = false
-pyramid.debug_notfound = false
-pyramid.debug_routematch = false
-pyramid.prevent_http_cache = false
-pyramid.default_locale_name = de
-
-
-# Mako specific
-
-api_url = ${API_URL}
-print_proxy_url = ${PRINT_PROXY_URL}
-#install_directory = ${current_directory}
-#host = ${host}
-#apache_base_path = ${apache_base_path}
-#apache_entry_path = ${apache_entry_path}
-print_temp_dir=${PRINT_TEMP_DIR}
-tomcat_port=${TOMCAT_PORT}
-
-
-###
-# wsgi server configuration
-###
-
-[server:main]
-use = egg:waitress#main
-host = 0.0.0.0
-port = 6543
-
-###
-# logging configuration
-# http://docs.pylonsproject.org/projects/pyramid/en/latest/narr/logging.html
-###
-
-[loggers]
-keys = root, print, sqlalchemy
-
-[logger_sqlalchemy]
-level = WARN
-handlers =
-qualname = sqlalchemy.engine
-# "level = INFO" logs SQL queries.
-# "level = DEBUG" logs SQL queries and results.
-# "level = WARN" logs neither.  (Recommended for production systems.)
-
-[handlers]
-keys = console
-
-[formatters]
-keys = generic
-
-[logger_root]
-level = WARN
-handlers = console
-
-[logger_print]
-level = WARN
-handlers =
-qualname = print
-
-[handler_console]
-class = StreamHandler
-args = (sys.stderr,)
-level = NOTSET
-formatter = generic
-
-[formatter_generic]
-format = %(asctime)s %(levelname)-5.5s [%(name)s][%(threadName)s] %(message)s
