+++ setup.py
@@ -8,29 +8,20 @@ CHANGES = open(os.path.join(here, 'CHANGES.txt')).read()
 
 requires = open(os.path.join(here, 'requirements.txt')).read().split('\n')
 
+
 setup(name='print3',
-      version='3.0.1',
+      version='0.0.1',
       description='Print service for geo.admin.ch',
       long_description=README + '\n\n' + CHANGES,
-      classifiers=[
-        "Programming Language :: Python",
-        "Framework :: Pyramid",
-        "Topic :: Internet :: WWW/HTTP",
-        "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
-        ],
+      classifiers=[],
+      keywords='',
       author='',
       author_email='',
-      url='',
-      keywords='web pyramid pylons',
-      packages=find_packages(),
-      package_data = {'print': ['locale/*/LC_MESSAGES/*.mo']},
+      license='MIT',
+      url='https://github.com/geoadmin/service-print',
+      packages=find_packages(exclude=['tests']),
+      package_dir={'print3': 'print3'},
       include_package_data=True,
       zip_safe=False,
-      install_requires=requires,
-      tests_require=requires,
-      test_suite="print",
-      entry_points="""\
-      [paste.app_factory]
-      main = print3:main
-      """,
+      test_suite='nose.collector',
       )
